+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.8] - 2026-02-13
+
+### Changed
+- Updated Guide page layout: merged Configuration and Private Repositories sections.
+- Replaced download buttons with simple links in Guide page.
+
 ## [1.0.7] - 2026-02-13
 
 ### Changed